pyproject.toml
[tool.poetry]
name = "neuro-nuggets"
version = "0.1.0"
description = ""
authors = ["Kevin Ding <dingkevin2005@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
server = "app.py"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-login = "^0.6.3"
pytest = "^8.2.0"
flask-sqlalchemy = "^3.1.1"
flask-socketio = "^5.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
